Address PR review: regex firmware matching, deduplicate vendor detection, gate UPnP source poll
